refactor(ui): extract HUD (stats panel, timer, readouts) to src/ui/hud.ts (R2.2)
Move the in-game HUD out of snowglider.ts into `src/ui/hud.ts`:
- `initializeGameStats(bestTime)` — best-time readout + Game Stats panel setup
- `initializeControlsToggle()` — Game Controls panel setup (both delegate to the
  shared `collapsible-panel` from R2.1, so the generic helper stays panel-agnostic)
- `updateStatsHud(result, pos, isInAir)` — per-frame speed (color-coded) /
  position / ground+technique readouts, lifted verbatim out of `updateSnowman()`
- `updateTimerDisplay(gameActive, startTime, bestTime)` — live run timer

Mechanical, behavior-preserving: the coordinator passes run state in as parameters
(it keeps owning `lastTechnique` and the landing-shake call in `updateSnowman`,
which move with the loop in a later step). No change to the published `window.*`
hooks or `publishGameGlobals()`. snowglider.ts drops ~93 net lines.

Verification: lint / typecheck / test (contract guard 44/44) / build green;
puppeteer 87/87 across 7/7 suites.
